requirements.txt
tkinter
pillow
lxml
pdfplumber
PyPDF2
pytesseract
tqdm
chardet
pandas
openpyxl